Clean up mycroft/audio/__init__.py
