+++ mycroft/audio/__init__.py
@@ -12,42 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import time
-
-from mycroft.util.signal import check_for_signal, create_signal
-
-
-def is_speaking():
-    """Determine if Text to Speech is occurring
-
-    Returns:
-        bool: True while still speaking
-    """
-    return check_for_signal("isSpeaking", -1)
-
-
-def wait_while_speaking():
-    """Pause as long as Text to Speech is still happening
-
-    Pause while Text to Speech is still happening.  This always pauses
-    briefly to ensure that any preceeding request to speak has time to
-    begin.
-    """
-    time.sleep(0.3)  # Wait briefly in for any queued speech to begin
-    while is_speaking():
-        time.sleep(0.1)
-
-
-def stop_speaking():
-    # TODO: Less hacky approach to this once Audio Manager is implemented
-    # Skills should only be able to stop speech they've initiated
-    from mycroft.messagebus.send import send
-    create_signal('stoppingTTS')
-    send('mycroft.audio.speech.stop')
-
-    # Block until stopped
-    while check_for_signal("isSpeaking", -1):
-        time.sleep(0.25)
-
-    # This consumes the signal
-    check_for_signal('stoppingTTS')
+from .utils import wait_while_speaking, is_speaking, stop_speaking
